refactor: move dropdown into wallet widget
