+++ apps/escrow/frontend/src/systems/Core/components/TopNav.tsx
@@ -1,5 +1,6 @@
 import {
   Box,
+  Button,
   Container,
   Flex,
   IconButton,
@@ -20,20 +21,8 @@ export function TopNav() {
   const { theme, toggleTheme } = useFuelTheme();
   const wallet = useWallet();
   const wallets = useWalletList();
-  const setCurrentWalletIndex = useSetAtom(walletIndexAtom);
 
-  const getWalletOptions = () => {
-    let walletOptions = [];
-    for (const [i, wallet] of wallets!.entries()) {
-      walletOptions.push(<option key={i} value={i}>{wallet?.address.slice(0, 4)}...{wallet?.address.slice(-4)}</option>);
-    }
-    return walletOptions;
-  }
 
-  const handleWalletChange = (event: ChangeEvent) => {
-    const currentWalletIndex = parseInt((event.target as HTMLInputElement).value);
-    setCurrentWalletIndex(currentWalletIndex);
-  };
 
   return (
     <Box css={{ borderBottom: "1px solid $gray5", background: "$gray3" }}>
@@ -41,9 +30,6 @@ export function TopNav() {
         <Flex css={{ py: "$8" }}>
           <Stack gap="$6" direction="row" css={{ flex: 1 }}></Stack>
         {wallet && <WalletWidget />}
-        <Dropdown onChange={handleWalletChange}>
-          {getWalletOptions()}
-        </Dropdown>
           <IconButton
             variant="link"
             aria-label="Theme Switcher"
